Correct length of flow update full
The offset of the match of a flow update full entry of
a multipart flow monitor reply is 24 not 32.

32, the length of a struct ofp_flow_update_full specified in the Section
7.3.5.17.2 Flow monitoring reply pf the OpenFlow 1.4.0 specification,
includes an 8 byte struct ofp_match. And the match portion of a flow update
full entry starts at the beginning of the match field and not the end of
it. That is at offset 32 - 8 = 24 and not offset 32.